requirements.txt
django
django-pandas
django-environ
pandas
plotly
django-import-export